+++ app/(anon)/interview/[id]/components/UserAudio.tsx
@@ -77,7 +77,7 @@ export default function UserAudio({
 
   return (
     <div
-      className={`w-full bg-white rounded-[8px] border border-[#E2E8F0] p-4 flex gap-[8px] items-center`}
+      className={`absolute bottom-[52px] w-[calc(100%-104px)] rounded-[8px] bg-white border border-[#E2E8F0] p-4 flex gap-[8px] items-center`}
     >
       <Mic width={16} height={16} />
       <div className="flex items-center text-[#1E293B] text-[12px] font-medium">
