+++ lib/avatar/audio.ts
