+++ lib/avatar/idle.ts
