+++ types/avatar.ts
